+++ src/main/kotlin/omok/model/Board.kt
@@ -1,4 +1,4 @@
-package omok
+package omok.model
 
 class Board {
     val board: List<MutableList<IntersectionState>> = List(15) { MutableList(15) { IntersectionState.EMPTY } }
@@ -14,19 +14,27 @@ class Board {
         val diagonalUp = checkLine(intersection, 6)
 
         if (verticalCheck || horizontalCheck || diagonalDown || diagonalUp) {
-            if (intersection.state == IntersectionState.BLACK)
+            if (intersection.state == IntersectionState.BLACK) {
                 return BoardState.BLACK_OMOK
-            if (intersection.state == IntersectionState.WHITE)
+            }
+            if (intersection.state == IntersectionState.WHITE) {
                 return BoardState.WHITE_OMOK
+            }
         }
         return BoardState.PLAYING
     }
 
-    private fun checkLine(intersection: Intersection, startDir: Int): Boolean {
+    private fun checkLine(
+        intersection: Intersection,
+        startDir: Int,
+    ): Boolean {
         return 1 + checkDirection(intersection, startDir) + checkDirection(intersection, startDir + 1) >= 5
     }
 
-    private fun checkDirection(intersection: Intersection, dir: Int): Int {
+    private fun checkDirection(
+        intersection: Intersection,
+        dir: Int,
+    ): Int {
         var curX = intersection.position.row.value
         var curY = intersection.position.column.value
         val curColor = intersection.state
@@ -35,19 +43,22 @@ class Board {
 
         var count = 0
         while (true) {
-            if (count >= 5)
+            if (count >= 5) {
                 return 5
+            }
 
             val nextX = curX + direction[dir].first
             val nextY = curY + direction[dir].second
 
-            if (nextX <= 0 || nextX >= 16 || nextY <= 0 || nextY >= 16)
+            if (nextX <= 0 || nextX >= 16 || nextY <= 0 || nextY >= 16) {
                 break
+            }
 
-            if (board[nextX][nextY] == curColor)
+            if (board[nextX][nextY] == curColor) {
                 count++
-            else
+            } else {
                 break
+            }
 
             curX = nextX
             curY = nextY
