+++ src/main/kotlin/omok/model/BoardState.kt
@@ -1,4 +1,4 @@
-package omok
+package omok.model
 
 enum class BoardState {
     WHITE_OMOK,
