+++ src/main/kotlin/omok/model/IntersectionState.kt
@@ -1,4 +1,4 @@
-package omok
+package omok.model
 
 enum class IntersectionState {
     EMPTY,
