+++ src/test/kotlin/omok/model/LineTest.kt
@@ -1,4 +1,4 @@
-package omok
+package omok.model
 
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
